Moved methods around TwiMLMessages to its own Class
