+++ src/main/java/com/twilio/survey/controllers/SurveyController.java
@@ -1,12 +1,10 @@
 package com.twilio.survey.controllers;
 
-import com.twilio.sdk.verbs.*;
 import com.twilio.survey.Server;
 import com.twilio.survey.models.Response;
 import com.twilio.survey.models.SurveyService;
 import com.twilio.survey.models.Survey;
 import com.twilio.survey.util.IncomingCall;
-import com.twilio.survey.util.Question;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
@@ -29,38 +27,16 @@ public class SurveyController {
     Survey existingSurvey = surveys.getSurvey(call.getFrom());
     if (existingSurvey == null) {
       Survey survey = surveys.createSurvey(call.getFrom());
-      return firstTwiMLQuestion(survey);
+      return TwiMLMessageFactory.firstTwiMLQuestion(survey);
     } else if (!existingSurvey.isDone()) {
       existingSurvey.appendResponse(new Response(call.getInput()));
       surveys.updateSurvey(existingSurvey);
       if (!existingSurvey.isDone()) {
-        return nextTwiMLQuestion(existingSurvey).toXML();
+        return TwiMLMessageFactory.nextTwiMLQuestion(existingSurvey).toXML();
       }
     }
-    return goodByeTwiMLMessage();
+    return TwiMLMessageFactory.goodByeTwiMLMessage();
   };
-
-  private static Object goodByeTwiMLMessage() throws TwiMLException {
-    TwiMLResponse twiml = new TwiMLResponse();
-    twiml.append(new Say("Your responses have been recorded. Thank you for your time!"));
-    return twiml.toXML();
-  }
-
-  private static Object firstTwiMLQuestion(Survey survey) throws TwiMLException, UnsupportedEncodingException {
-    TwiMLResponse response = new TwiMLResponse();
-    response.append(new Say("Thanks for taking our survey."));
-    return nextTwiMLQuestion(survey, response).toXML();
-  }
-
-  private static Verb nextTwiMLQuestion(Survey survey) throws TwiMLException, UnsupportedEncodingException {
-    return nextTwiMLQuestion(survey, null);
-  }
-
-  protected static TwiMLResponse nextTwiMLQuestion(Survey survey, TwiMLResponse twiml) throws TwiMLException,
-      UnsupportedEncodingException {
-    Question question = Server.config.getQuestions()[survey.getIndex()];
-    return buildQuestionTwiML(survey, question, twiml);
-  }
 
 
   // Results accessor route
@@ -91,64 +67,6 @@ public class SurveyController {
     };
 
   // Helper methods
-
-  private static TwiMLResponse buildQuestionTwiML(Survey survey, Question question, TwiMLResponse twiml) throws TwiMLException, UnsupportedEncodingException {
-    TwiMLResponse response = (twiml != null) ? twiml : new TwiMLResponse();
-    Say say = new Say(question.getText());
-    response.append(say);
-    // Depending on the question type, create different TwiML verbs.
-    switch (question.getType()) {
-      case "text":
-        response = appendTextQuestion(survey, response);
-        break;
-      case "boolean":
-        response = appendBooleanQuestion(response);
-        break;
-      case "number":
-        response = appendNumberQuestion(response);
-        break;
-    }
-    return response;
-  }
-
-  private static TwiMLResponse appendNumberQuestion(TwiMLResponse twiml) throws TwiMLException {
-    Say numInstructions = new Say("Enter the number on your keypad, followed by the #.");
-    twiml.append(numInstructions);
-    Gather numberGather = new Gather();
-    // Listen until a user presses "#"
-    numberGather.setFinishOnKey("#");
-    twiml.append(numberGather);
-
-    return twiml;
-  }
-
-  private static TwiMLResponse appendBooleanQuestion(TwiMLResponse twiml) throws TwiMLException {
-    Say boolInstructions =
-        new Say("Press 0 to respond 'No,' and press any other number to respond 'Yes.'");
-    twiml.append(boolInstructions);
-    Gather booleanGather = new Gather();
-    // Listen only for one digit.
-    booleanGather.setNumDigits(1);
-    twiml.append(booleanGather);
-
-    return twiml;
-  }
-
-  private static TwiMLResponse appendTextQuestion(Survey survey, TwiMLResponse twiml) throws TwiMLException, UnsupportedEncodingException {
-    Say textInstructions =
-        new Say(
-            "Your response will be recorded after the tone. Once you have finished recording, press the #.");
-    twiml.append(textInstructions);
-    Record text = new Record();
-    text.setFinishOnKey("#");
-    // Use the Transcription route to receive the text of a voice response.
-    text.setTranscribe(true);
-    text.setTranscribeCallback("/interview/" + urlEncode(survey.getPhone()) + "/transcribe/"
-        + survey.getIndex());
-    twiml.append(text);
-
-    return twiml;
-  }
 
   // Spark has no built-in body parser, so let's roll our own.
   public static Map<String, String> parseBody(String body) throws UnsupportedEncodingException {
